server2 -> server + extract classes to separate files
