+++ lib/mayu/commands/dev.rb
@@ -2,7 +2,7 @@
 # frozen_string_literal: true
 
 require_relative "base"
-require_relative "../server2"
+require_relative "../server"
 
 module Mayu
   module Commands
@@ -12,7 +12,7 @@ module Mayu
       sig { params(argv: T::Array[String]).void }
       def call(argv)
         Configuration.log_config(configuration)
-        Server2.start_dev(configuration)
+        Server.start_dev(configuration)
       end
     end
   end
